refactor: seperate VoiceTimer component.
